+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.2 (in development)
 
  * Remove JSON gem dependency since no longer supporting Ruby 1.8.7 anyhow.
+ * Add `EmojiData.find_by_str` convenience method to match on a string.
  * Make default `EmojiChar.to_s()` the same as `EmojiChar.char()`